+++ content-off_canvas_mobile.php
@@ -2,7 +2,7 @@
 /**
  * The template that supplies Off-Canvas mobile support to WP-Forge
  * @since WP-Forge 5.5.2.2
- * @version 6.2.2
+ * @version 6.2.3
  */
 ?>
 <div class="off-canvas-wrapper">