+++ functions.php
@@ -6,7 +6,7 @@
  * Others are attached to action and filter hooks in WordPress to change core functionality. Any new functions will
  * be added at the end of the file. This will allow everyone to keep track of what has been added.
  * @since WP-Forge 5.5.1.7
- * @version 6.2.2
+ * @version 6.2.3
  */
 
 /**
@@ -159,19 +159,19 @@ if ( ! function_exists( 'wpforge_google_fonts' ) ) {
  */
 function wpforge_scripts() {
 	global $wp_styles;
-  wp_enqueue_style('wpforge_fonts', get_template_directory_uri() . '/fonts/fonts.css','', '6.2.2' );
+  wp_enqueue_style('wpforge_fonts', get_template_directory_uri() . '/fonts/fonts.css','', '6.2.3' );
 
 	if( get_theme_mod( 'wpforge_select_css' ) == 'flex') {
-			wp_enqueue_style('wpforge_foundation', get_template_directory_uri() . '/css/foundation-flex.css','', '6.2.2' );
+			wp_enqueue_style('wpforge_foundation', get_template_directory_uri() . '/css/foundation-flex.css','', '6.2.3' );
 	} else {
-			wp_enqueue_style('wpforge_foundation', get_template_directory_uri() . '/css/foundation.css','', '6.2.2' );
+			wp_enqueue_style('wpforge_foundation', get_template_directory_uri() . '/css/foundation.css','', '6.2.3' );
 	}
 
-  wp_enqueue_style('wpforge_motion_ui', get_template_directory_uri() . '/css/motion-ui.css','', '1.2.2' );
-  wp_enqueue_style('wpforge', get_stylesheet_uri(),'','6.2.1.2' );
-	wp_enqueue_script('wpforge_what_input', get_template_directory_uri() . '/js/what-input.js', array('jquery'),'6.2.2', true);
-	wp_enqueue_script('wpforge_foundation', get_template_directory_uri() . '/js/foundation.js', array('jquery'),'6.2.2', true);
-	wp_enqueue_script('wpforge_functions', get_template_directory_uri() . '/js/theme-functions.js', array('jquery'),'6.2.2', true);
+  	wp_enqueue_style('wpforge_motion_ui', get_template_directory_uri() . '/css/motion-ui.css','', '1.2.2' );
+  	wp_enqueue_style('wpforge', get_stylesheet_uri(),'','6.2.1.2' );
+	wp_enqueue_script('wpforge_what_input', get_template_directory_uri() . '/js/what-input.js', array('jquery'),'6.2.3', true);
+	wp_enqueue_script('wpforge_foundation', get_template_directory_uri() . '/js/foundation.js', array('jquery'),'6.2.3', true);
+	wp_enqueue_script('wpforge_functions', get_template_directory_uri() . '/js/theme-functions.js', array('jquery'),'6.2.3', true);
 }
 add_action( 'wp_enqueue_scripts', 'wpforge_scripts', 0);
 /**
@@ -180,7 +180,7 @@ add_action( 'wp_enqueue_scripts', 'wpforge_scripts', 0);
  */
 if ( ! function_exists( 'wpforge_prepare_foundation' ) ) {
 	function wpforge_prepare_foundation() {
-		wp_enqueue_script ('wpforge_load_foundation', get_template_directory_uri() . '/js/app.js', array('wpforge_foundation'), '6.2.2', true);
+		wp_enqueue_script ('wpforge_load_foundation', get_template_directory_uri() . '/js/app.js', array('wpforge_foundation'), '6.2.3', true);
 	}
 	add_action( 'wp_enqueue_scripts', 'wpforge_prepare_foundation', 999);
 }