+++ js/app.js
@@ -1,5 +1,5 @@
 /**
  * @since WP-Forge 6.1.1
- * @version 6.2.2
+ * @version 6.2.3
  */
 jQuery(document).foundation();